Psuedo Code part 3
Added some more real code to the pseudo code that will assist in writing
the program. Also changed the md file to a txt file. Made a for loop
inside a for loop that will cycle through the different dates of each
csv file and choose only the most current of each company.